don't flush redis cache in production

+++ lib/flushCache.coffee
@@ -1,3 +1,6 @@
 {exec} = require 'child_process'
-module.exports = (cb)->
-  exec "redis-cli FLUSHALL", cb
+module.exports = (cb) ->
+  unless NODE_ENV is 'production'
+    exec "redis-cli FLUSHALL", cb
+  else
+    cb()
